refactor: move image name helper to utils
