+++ internal/shared/utils/image_test.go
